requirements.txt
streamlit==1.32.0
langchain==0.1.12
python-dotenv==1.0.1
google-generativeai==0.3.2
numpy==1.26.4
--only-binary :all: 